Controller Tests Done

+++ capstone-pm/src/main/java/ideas/capstone_pm/service/CartService.java
@@ -1,8 +1,7 @@
 package ideas.capstone_pm.service;
 
 import ideas.capstone_pm.dto.AddToCartDTO;
-import ideas.capstone_pm.dto.CartProjection;
-import ideas.capstone_pm.dto.DeleteCartDTO;
+import ideas.capstone_pm.projection.CartProjection;
 import ideas.capstone_pm.entity.ApplicationUser;
 import ideas.capstone_pm.entity.Cart;
 import ideas.capstone_pm.entity.Fund;
@@ -12,14 +11,12 @@ import ideas.capstone_pm.repository.CartRepository;
 import ideas.capstone_pm.repository.TransactionRepository;
 import ideas.capstone_pm.util.CartServiceUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class CartService {
@@ -39,57 +36,62 @@ public class CartService {
         }
     }
 
-    public List<CartProjection> getCartsByUser(ApplicationUser user) {
+    public List<CartProjection> getCartsByUser(Integer userId) {
         try {
+            ApplicationUser user = new ApplicationUser();
+            user.setUserId(userId);
             return cartRepository.findByUser(user);
         } catch (Exception e) {
-            throw new CartNotFoundException();
+            throw new CartNotFoundException("Cart Not Found");
         }
     }
 
     public void deleteCartByCartId(Integer cartId) {
-        try {
-            Optional<Cart> cart = cartRepository.findById(cartId);
-            if (cart.isPresent()) {
-                cartRepository.deleteById(cartId);
-            } else {
-                throw new RuntimeException("Cart not found with ID: " + cartId);
-            }
-        } catch (Exception e) {
-            throw new RuntimeException("An error occurred while deleting the cart.");
+        Optional<Cart> cart = cartRepository.findById(cartId);
+        if (cart.isPresent()) {
+            cartRepository.deleteById(cartId);
+        } else {
+            throw new CartNotFoundException("Cart not found");
         }
     }
 
     public double computeTotalAmountInCart(ApplicationUser user) {
-        try {
-            List<CartProjection> userCarts = cartRepository.findByUser(user);
-            return userCarts.stream().mapToDouble(CartProjection::getPlannedInvestment).sum();
-        } catch (Exception e) {
-            throw new RuntimeException("An error occurred while computing the total amount in the cart.");
-        }
+        List<CartProjection> userCarts = cartRepository.findByUser(user);
+        return userCarts.stream()
+                .mapToDouble(CartProjection::getPlannedInvestment)
+                .sum();
     }
-
 
     public void processCart(ApplicationUser user) {
+        if (user == null) {
+            throw new IllegalArgumentException("User cannot be null");
+        }
+
+        List<CartProjection> cartItems = cartRepository.findByUser(user);
+        if (cartItems.isEmpty()) {
+            return;
+        }
+
         try {
-            List<CartProjection> cartItems = cartRepository.findByUser(user);
-
-            for (CartProjection cartItem : cartItems) {
-                Transaction transaction = new Transaction();
-                Fund fund = new Fund();
-                fund.setFundId(cartItem.getFund().getFundId());
-
-                transaction.setFund(fund);
-                transaction.setUser(user);
-                transaction.setTransactionInitialInvestment(cartItem.getPlannedInvestment());
-                transaction.setTransactionDate(new Date());
-                transactionRepository.save(transaction);
-            }
-
+            List<Transaction> transactions = createTransactions(cartItems, user);
+            transactionRepository.saveAll(transactions);
             cartRepository.deleteByUser(user);
         } catch (Exception e) {
-            throw new RuntimeException("An error occurred while processing the cart.");
+            throw new RuntimeException("An error occurred while processing the cart.", e);
         }
     }
 
+    private List<Transaction> createTransactions(List<CartProjection> cartItems, ApplicationUser user) {
+        return cartItems.stream().map(cartItem -> {
+            Transaction transaction = new Transaction();
+            Fund fund = new Fund();
+            fund.setFundId(cartItem.getFund().getFundId());
+
+            transaction.setFund(fund);
+            transaction.setUser(user);
+            transaction.setTransactionInitialInvestment(cartItem.getPlannedInvestment());
+            transaction.setTransactionDate(new Date());
+            return transaction;
+        }).collect(Collectors.toList());
+    }
 }
